Updated paginate() to use the original query to obtain a COUNT(), but limit total to 2000 because that is all SOQL's OFFSET supports.  Prior to this change, paginate() was useless if your query had a WHERE clause.
Updated select() in SOQLConnection to return *all* results from SalesForce (by continually calling next()) rather than just up to 2000. Using the method can certaily lead to long load times and high memory usage, however I beleive it's the responsiblity of the caller to use LIMIT or cursor() (which has been implemented in this commit as well) in order to reduce memory usage.   select() is defined by Eloquent to return *all* results so I believe that should be the behavior of this SOQL driver as well.

Implmented the cursor() method to support yield-ing over a result set.  It's important to note that with a typical RDBM driver for Eloquent, cursor() would allow you to store one row in memory at any given time.  However due to the nature of SOQL and it being a remote API with quota restrictions, it's unreasonable to make API calls row-by-row.  So cursor() that has been implemented will still store up to the SOQL-max of 2000 records at once, however it will yeild them one by one and call next() transparently in the background for the Eloquent user.